pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["sanjin-lebesgue <sanjin@lebesgue.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
numba = "0.57.1"
shap = "^0.42.1"
bcrypt = "^4.0.1"
black = "^23.7.0"
loguru = "^0.7.0"
matplotlib = "^3.7.2"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
scikit-learn = "^1.3.0"
streamlit = "^1.25.0"
streamlit-authenticator = "^0.2.2"
streamlit-option-menu = "^0.3.6"
tqdm = "^4.65.0"
ipython = "^8.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
